+++ R/manyenviron-agreements.R
@@ -1,16 +1,13 @@
 #' agreements database documentation
 #'
 #' @format The agreements database is a list that contains the
-#' following 5 datasets: IEADB, GNEVAR, ECOLEX, CIESIN, HEIDI.
+#' following 6 datasets: IEADB, ECOLEX, CIESIN, HEIDI, HUGGO, AIGGO.
 #' For more information and references to each of the datasets used,
 #' please use the `data_source()` and `data_contrast()` functions.
 #'\describe{
 #' \item{IEADB: }{A dataset with 3666 observations and the following
 #' 10 variables: manyID, Title, Beg, DocType, AgreementType, Signature,
 #' Force, Lineage, treatyID, ieadbID.}
-#' \item{GNEVAR: }{A dataset with 7273 observations and the following
-#' 14 variables: manyID, Title, Beg, End, DocType, AgreementType, GeogArea,
-#' Signature, Force, Lineage, accessionC, accessionP, treatyID, gnevarID.}
 #' \item{ECOLEX: }{A dataset with 2174 observations and the following
 #' 10 variables: manyID, Title, Beg, DocType, GeogArea, Signature, Force,
 #' Lineage, treatyID, ecolexID.}
@@ -18,7 +15,19 @@
 #' 7 variables: manyID, Title, Beg, Signature, Force, Lineage, treatyID.}
 #' \item{HEIDI: }{A dataset with 2280 observations and the following
 #' 7 variables: manyID, Title, Beg, Signature, Lineage, treatyID, heidiID.}
+#' \item{HUGGO: }{A dataset with 7330 observations and the following
+#' 49 variables: manyID, treatyID, Title, Beg, End, Signature, Force, DocType,
+#' AgreementType, GeogArea, gnevarID, verified, gengID, ecolexID, ieaID,
+#' DocValidUntilDate, Notes, Download, MEA_type, Ambit, Region, subject_ecolex,
+#' subject_iea, Keywords, Lineage, Sequence, AdoptedIn, Languages, Appendices,
+#' Depository, DepositoryURL, Published, Abstract, Website1, Website2,
+#' Secretariat, SecretariatURL, UNEP, Supersedes, References, EnabledBy,
+#' AmendedBy, Lit, Data, Coded, TreatyText, SourceText, Text, AbstractText.}
+#' \item{AIGGO: }{A dataset with 7330 observations and the following
+#' 6 variables: manyID, Lineage, accessionC, accessionP, termination_type,
+#' termination_date.}
 #' }
+
 #'
 #' @details
 #' ``` {r, echo = FALSE, warning = FALSE}