+++ data-raw/memberships/HUGGO_MEM/prepare-HUGGO_MEM.R
@@ -1,0 +1,118 @@
+# HUGGO_MEM Preparation Script
+
+# This is a template for importing, cleaning, and exporting data
+# ready for the many packages universe.
+
+# Stage one: Collecting data
+HUGGO_MEM <- readr::read_csv("data-raw/memberships/HUGGO_MEM/gnevar.csv")
+
+# Stage two: Correcting data
+# In this stage you will want to correct the variable names and
+# formats of the 'HUGGO_MEM' object until the object created
+# below (in stage three) passes all the tests.
+HUGGO_MEM <- as_tibble(HUGGO_MEM) %>%
+  janitor::remove_empty(which = "cols") %>%
+  dplyr::mutate(gengID = GENG,
+                ecolexID = ECOLEX,
+                ieaID = IEA,
+                CountryID = Country,
+                CountryRat = messydates::as_messydate(Approval),
+                CountrySignature = messydates::as_messydate(Signature),
+                CountryWithdrawal = messydates::as_messydate(Withdrawal1),
+                CountryWithdrawal2 = messydates::as_messydate(Withdrawal2),
+                Signature = messydates::as_messydate(DocSign),
+                Force = messydates::as_messydate(DocForce),
+                Term = messydates::as_messydate(DocEnd),
+                CountryForce = messydates::as_messydate(InForce1),
+                CountryForce2 = messydates::as_messydate(InForce2),
+                CountryForce3 = messydates::as_messydate(InForce3),
+                verified = case_when(X == "%" ~ "verified",
+                                     X == "?" ~ "not verified"),
+                Title = manypkgs::standardise_titles(Title),
+                ProvisionalApp = messydates::as_messydate(ProvisionalApp),
+                Deposit = messydates::as_messydate(Deposit),
+                Beg = dplyr::coalesce(Signature, CountryRat, CountryForce),
+                End = dplyr::coalesce(Term, CountryWithdrawal)) %>%
+  dplyr::select(CountryID, Title, Beg, End, Signature, CountrySignature,
+                CountryRat, Force, CountryForce, CountryForce2, CountryForce3,
+                Term, CountryWithdrawal, CountryWithdrawal2,
+                gengID, ecolexID, ieaID, Comments, Deposit, obsolete,
+                ProvisionalApp, Reservation, verified) %>%
+  dplyr::arrange(Beg) %>% 
+  dplyr::distinct()
+
+# Add treatyID column
+HUGGO_MEM$treatyID <- manypkgs::code_agreements(HUGGO_MEM,
+                                                HUGGO_MEM$Title,
+                                                HUGGO_MEM$Beg)
+
+# Add MEA edges data (MGENG dataset)
+# For some more information about the variables and codes,
+# please see the documentation in the data-raw folder.
+MEA_edges <- readr::read_delim("data-raw/agreements/HUGGO/MEA.Edges v1.0.csv", 
+                               delim = ";", escape_double = FALSE, trim_ws = TRUE)
+# Let's wrangle some variables to keep consistent.
+names(MEA_edges) <- gsub("\\.", "", names(MEA_edges))
+names(MEA_edges) <- gsub("^Membership", "", names(MEA_edges))
+names(MEA_edges) <- gsub("^dateOf", "", names(MEA_edges))
+MEA_edges <- as_tibble(MEA_edges) %>%
+  dplyr::select(-'1') %>%
+  janitor::remove_empty(which = "cols") %>%
+  manydata::transmutate(gengID = GENGID,
+                        CountryID = Country,
+                        CountrySignature = messydates::as_messydate(MembSignx),
+                        CountryRat = messydates::as_messydate(MembRatx),
+                        CountryForce_ecolex = messydates::as_messydate(MembForcex),
+                        CountryForce_iea = messydates::as_messydate(MembForcey),
+                        CountryForce2 = messydates::as_messydate(EntryintoForce2),
+                        CountryWithdrawal = messydates::as_messydate(Withdrawal),
+                        ProvisionalApp = messydates::as_messydate(ProvisionalApplication),
+                        Succession = messydates::as_messydate(Succession),
+                        Deposit = messydates::as_messydate(DepositofInstrument),
+                        CooperatingNonparty = messydates::as_messydate(CooperatingNonparty),
+                        DefiniteSignature = messydates::as_messydate(DefiniteSignature),
+                        Consent = messydates::as_messydate(ConsentToBeBound),
+                        Acceptance = messydates::as_messydate(AcceptanceApproval),
+                        Accession = messydates::as_messydate(AccessionApprobation),
+                        Consolidation = messydates::as_messydate(Consolidation),
+                        Acceptance = messydates::as_messydate(AccessionApprobation2),
+                        CountryWithdrawal = messydates::as_messydate(Withdrawal)) %>% 
+  dplyr::distinct()
+
+# Add titles and ID variables from HUGGO agreements data
+agreements <- manyenviron::agreements$HUGGO %>%
+  dplyr::select(c(gengID, Title, Beg, End, Signature, Force, ecolexID, ieaID))
+MEA_edges <- dplyr::inner_join(MEA_edges, agreements, by = "gengID") %>% 
+  dplyr::distinct()
+
+# Add treatyID column
+MEA_edges$treatyID <- manypkgs::code_agreements(MEA_edges,
+                                                MEA_edges$Title,
+                                                MEA_edges$Beg)
+
+# Join Data 
+HUGGO_MEM <- dplyr::full_join(HUGGO_MEM, MEA_edges) %>% 
+  dplyr::distinct()
+
+# Add manyID column
+manyID <- manypkgs::condense_agreements(var = HUGGO_MEM$treatyID)
+HUGGO_MEM <- dplyr::left_join(HUGGO_MEM, manyID, by = "treatyID")
+
+# Reorder variables
+HUGGO_MEM <- dplyr::relocate(HUGGO_MEM, c("manyID", "treatyID", "CountryID", "Title",
+                                          "Beg", "End", "Signature", "Force"))
+
+# Stage three: Connecting data
+# Next run the following line to make HUGGO_MEM available
+# within the package.
+manypkgs::export_data(HUGGO_MEM, database = "memberships", URL = "NA")
+# This function also does two additional things.
+# First, it creates a set of tests for this object to ensure
+# adherence to certain standards. You can hit Cmd-Shift-T (Mac)
+# or Ctrl-Shift-T (Windows) to run these tests locally at any point.
+# Any test failures should be pretty self-explanatory and may require
+# you to return to stage two and further clean, standardise, or wrangle
+# your data into the expected format.
+# Second, it also creates a documentation file for you to fill in.
+# Please make sure that you cite any sources appropriately and fill
+#in as much detail about the variables etc as possible.